Moves peer count format function to utils to fix 'this' issue
